+++ CHANGELOG.md
@@ -7,6 +7,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2026-02-26
+
+### Added
+
+- `ExactSizeIterator` implementation for `InvalidIterator`
+- `TestIterator` - iterator that can not be iterated over, but has an arbitrary size hint
+
+### Removed
+
+- Unecessary sealing, and `sealed` crate.
+
 ## [0.4.1] - 2026-01-30
 
 ### Added